backend/requirements.txt
django>=6.0,<7.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.0,<5.0
boto3>=1.35,<2.0
